♻️  Apply renaming in all components
