+++ packages/applications/ssr/src/components/pages/abandon/détails/EtapesAbandon.tsx
@@ -8,7 +8,7 @@ import { Routes } from '@potentiel-applications/routes';
 import { Iso8601DateTime } from '@potentiel-libraries/iso8601-datetime';
 
 import { Timeline, TimelineProps } from '@/components/organisms/Timeline';
-import { InputDownload } from '@/components/atoms/form/InputDownload';
+import { DownloadDocument } from '@/components/atoms/form/DownloadDocument';
 
 import { StatutPreuveRecandidatureBadgeProps } from './PreuveRecandidatureStatutBadge';
 
@@ -97,12 +97,13 @@ export const EtapesAbandon: FC<EtapesAbandonProps> = ({
       content: (
         <>
           {accord.réponseSignée && (
-            <InputDownload
-              details=""
-              ariaLabel="Télécharger la pièce justificative pour l'accord de la demande d'abandon"
+            <DownloadDocument
               label="Télécharger la pièce justificative"
+              details=""
               linkProps={{
                 href: Routes.Document.télécharger(accord.réponseSignée),
+                'aria-label':
+                  "Télécharger la pièce justificative pour l'accord de la demande d'abandon",
               }}
               className="mb-0 pb-0"
             />
@@ -122,12 +123,13 @@ export const EtapesAbandon: FC<EtapesAbandonProps> = ({
       content: (
         <>
           {rejet.réponseSignée && (
-            <InputDownload
-              details=""
-              ariaLabel="Télécharger la pièce justificative pour le rejet de la demande d'abandon"
+            <DownloadDocument
               label="Télécharger la pièce justificative"
+              details=""
               linkProps={{
                 href: Routes.Document.télécharger(rejet.réponseSignée),
+                'aria-label':
+                  "Télécharger la pièce justificative pour le rejet de la demande d'abandon",
               }}
               className="mb-0 pb-0"
             />
@@ -160,12 +162,13 @@ export const EtapesAbandon: FC<EtapesAbandonProps> = ({
       content: (
         <>
           {confirmation.réponseSignée && (
-            <InputDownload
-              details=""
-              ariaLabel="Télécharger la pièce justificative pour la confirmation de la demande d'abandon"
+            <DownloadDocument
               label="Télécharger la pièce justificative"
+              details=""
               linkProps={{
                 href: Routes.Document.télécharger(confirmation.réponseSignée),
+                'aria-label':
+                  "Télécharger la pièce justificative pour la confirmation de la demande d'abandon",
               }}
               className="mb-0 pb-0"
             />
@@ -193,12 +196,12 @@ export const EtapesAbandon: FC<EtapesAbandonProps> = ({
           </blockquote>
         </div>
         {justificatifDemande && (
-          <InputDownload
-            details=""
-            ariaLabel="Télécharger la pièce justificative pour la demande d'abandon"
+          <DownloadDocument
             label="Télécharger la pièce justificative"
+            details=""
             linkProps={{
               href: Routes.Document.télécharger(justificatifDemande),
+              'aria-label': "Télécharger la pièce justificative pour la demande d'abandon",
             }}
             className="mb-0 pb-0"
           />
