+++ packages/applications/ssr/src/components/pages/abandon/détails/demanderConfirmation/DemanderConfirmationAbandon.tsx
@@ -7,7 +7,7 @@ import { useState } from 'react';
 import { Routes } from '@potentiel-applications/routes';
 
 import { ModalWithForm } from '@/components/molecules/ModalWithForm';
-import { InputDownload } from '@/components/atoms/form/InputDownload';
+import { DownloadDocument } from '@/components/atoms/form/DownloadDocument';
 import { UploadDocument } from '@/components/atoms/form/UploadDocument';
 
 import { demanderConfirmationAbandonAction } from './demanderConfirmation.action';
@@ -58,10 +58,10 @@ export const DemanderConfirmationAbandon = ({
                 className="mb-8"
               />
 
-              <InputDownload
-                ariaLabel={`Télécharger le modèle de réponse pour demander la confirmation de la demande d'abandon `}
+              <DownloadDocument
                 linkProps={{
                   href: Routes.Abandon.téléchargerModèleRéponse(identifiantProjet),
+                  'aria-label': `Télécharger le modèle de réponse pour demander la confirmation de la demande d'abandon`,
                 }}
                 details="docx"
                 label="Télécharger le modèle de réponse"
